Create main script files
NotBatch will contain all of the important module updating and
versioning logic, and the Install module will be used to install all the
other modules in the config. On first install, it should add update
hooks into Microsoft.PowerShell_profile.ps1 and create the profile
script if it doesn't exist.